requirements.txt
Flask==0.12.4
Flask-Migrate==1.8.0
Flask-Negotiate==0.1.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
SQLAlchemy==1.0.13
Werkzeug==0.11.10
alembic==0.8.6
click==6.6
itsdangerous==0.24
psycopg2==2.6.1
python-editor==1.0.1
Flask-LogConfig==0.4.2
requests==2.20.0
amqp==1.4.9
anyjson==0.3.3
kombu==3.0.35
pyproj==1.9.5.1
pycryptodome==3.7.0
jsonschema==2.5.1
